deployment/requirements.txt
# Production ML Pipeline Dependencies
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
tensorflow==2.14.0
opencv-python-headless==4.8.0
pypdf2==3.0.1
plotly==5.18.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Core ML libraries
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.4.2
scipy==1.13.0

# Build dependencies
setuptools>=69.0.0
wheel>=0.43.0
Cython>=3.0.10
